generateChoices {shinyQnD}	R Documentation

Dependent choices for selectInput

Description

This function will generate the available options of a shiny::selectInput based on what is already selected. Often generateChoices_generateTableData can be used instead to generate the selectInput choices at the same time.

Usage

generateChoices(input ,data, columnNames)

Arguments

session
simply pass your input vector from your shiny app

data
dataframe (this will be used to generate what choices are available)

columnNames
collection of column names to base the choices on.

Author(s)

Bryen Moberg

See Also

generateChoices_generateData, generateSelectInput, resetSelectInputs

Examples

library(shiny)
library(ShinyQnD)
library(shinydashboard)

ui <- function(){
  dashboardPage(
    dashboardHeader(),
    dashboardSidebar(
      uiOutput("selectOut"),
      actionButton("reset","Reset")),
    dashboardBody(tableOutput("tableOut"))
  )
}

server <- function(input, output, session){

  ##### Server Code #####

  data <- reactive({
    read.csv("data.csv", stringsAsFactors = F)
  })

  resetSelections <- observeEvent(input$reset,{
    resetSelectInputs(c("COLOR","SIZE"))
  })

  output$selectOut <- renderUI({
    generateSelectInput(c("COLOR","SIZE"))
  })

  output$tableOut <- renderTable({
    tableData()
    })

  tableData <- reactive({
    generateChoices(data(),c("COLOR", "SIZE"))
    data <- generateTableData(input, data(),c("COLOR", "SIZE"))
    data
  })

}

shinyApp(ui <- ui, server <- server)

## The function is currently defined as
function (input, data, columnNames)
{
    ids <- columnNames
    for (id in ids) {
        filterData <- data
        for (filterid in ids[-grep(id, ids)]) {
            if (!is.null(input[[filterid]]) && input[[filterid]] !=
                "All") {
                filterData <- filterData[which(filterData[filterid] ==
                  input[[filterid]]), ]
            }
        }
        updateSelectInput(session, id, choices = c("All", unique(filterData[,
            id])), selected = input[[id]])
    }
  }

[Package shinyQnD version 0.0.2 Index]