generateTableData {shinyQnD}	R Documentation

Generating dataframe based on shiny::selectInputs

Description

This function generates a dataframe based on the given data and selectInput ids.

Usage

generateTableData(input, data, checkInputs)

Arguments

input
shiny input vector

data
data to use for dataframe

checkInputs
the ids of selectInputs to use

Value

data
returns a dataframe filtered by the value of the given selectInputs

Author(s)

Bryen Moberg

See Also

generateChoices_generateTableData generateSelectInput

Examples

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

library(shiny)
 ui <- fluidPage(actionButton("reset","Reset")
 selectInput("COLOR", "Color", choices = c("All", "Other")))

 server <- function(input, output, session){

 resetSelections <- observeEvent(input$reset,{
 resetSelectInputs(c("COLOR"))
 })
}
shinyApp(server = server, ui  = ui)

## The function is currently defined as
resetSelectInputs <- function(session,inputNames){
  for(name in inputNames){
    updateVarSelectInput(session, name, selected = "All")
  }
}

## The function is currently defined as
function (input, data, checkInputs)
{
    for (column in checkInputs) {
        if (!is.null(input[[column]]) && input[[column]] != "All") {
            data <- data[which(data[column] == input[[column]]),
                ]
        }
    }
    data
  }

[Package shinyQnD version 0.0.2 Index]